app/requirements.txt
Django==3.2.18
django-split-settings==1.2.0
python-dotenv==1.0.0
psycopg2==2.9.5
uwsgi==2.0.21
pydantic==1.10.5
djangorestframework==3.14.0
python-dateutil==2.8.2
stripe==5.2.0

PyJWT==2.6.0
cryptography==38.0.4
